+++ src/app/static/src/store/getters.ts
@@ -1,7 +1,7 @@
 import dayjs from "dayjs";
 import { RootState } from "@/store/state";
 import { Data } from "@/types";
-import { numericFormatter } from "@/components/parameters/phasesUtils";
+import { numericFormatter } from "../utils/formatter";
 
 // We hard-code this parameter sent to cometr to give a forecast of 2 years
 export const forecastDays = 730;
